Revert "fix(core): u16→u32 for extra_data and witness length in serialization"
This reverts commit 37a14ae6c8314fa623539f7c1b33b8f83da5dfce.

+++ crates/core/src/transaction/output.rs
@@ -165,7 +165,7 @@ pub struct Output {
     /// Lock until height (0 for normal, >0 for bonds)
     pub lock_until: BlockHeight,
     /// Extensible data for future output types (empty for Normal/Bond).
-    /// Interpretation depends on output_type. Max 512 KB (era 0), grows per era.
+    /// Interpretation depends on output_type. Max 512 bytes.
     #[serde(default)]
     pub extra_data: Vec<u8>,
 }
@@ -629,9 +629,8 @@ impl Output {
         bytes.extend_from_slice(&self.amount.to_le_bytes());
         bytes.extend_from_slice(self.pubkey_hash.as_bytes());
         bytes.extend_from_slice(&self.lock_until.to_le_bytes());
-        // extra_data: length-prefixed (u32 LE) + raw bytes
-        // u32 supports up to 4GB, well above MAX_EXTRA_DATA_SIZE (512KB+)
-        bytes.extend_from_slice(&(self.extra_data.len() as u32).to_le_bytes());
+        // extra_data: length-prefixed (u16 LE) + raw bytes
+        bytes.extend_from_slice(&(self.extra_data.len() as u16).to_le_bytes());
         bytes.extend_from_slice(&self.extra_data);
         bytes
     }
